pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iodd-manager"
version = "2.0.1"
description = "IODD Manager - Industrial automation tool for managing IO-Link Device Description files"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "ME-Catalyst", email = "noreply@me-catalyst.com"}
]
maintainers = [
    {name = "ME-Catalyst", email = "noreply@me-catalyst.com"}
]
keywords = ["iodd", "io-link", "industrial-automation", "iot", "fastapi", "device-description"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Hardware",
    "Topic :: Scientific/Engineering",
    "Framework :: FastAPI",
]

dependencies = [
    "lxml>=4.9.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.0",
    "python-dotenv>=1.0.0",
    "xmlschema>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "pylint>=2.17.0",
    "mypy>=1.4.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.1.0",
]
advanced = [
    "aiofiles>=23.0.0",
    "redis>=4.6.0",
    "celery>=5.3.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
]
security = [
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]
all = [
    "iodd-manager[dev,docs,advanced,security]",
]

[project.urls]
Homepage = "https://github.com/ME-Catalyst/iodd-manager"
Documentation = "https://github.com/ME-Catalyst/iodd-manager/blob/main/docs/INDEX.md"
Repository = "https://github.com/ME-Catalyst/iodd-manager"
Issues = "https://github.com/ME-Catalyst/iodd-manager/issues"
Changelog = "https://github.com/ME-Catalyst/iodd-manager/blob/main/CHANGELOG.md"

[project.scripts]
greenstack = "src.greenstack:main"
greenstack-api = "src.api:main"

[tool.setuptools]
packages = ["src", "src.routes", "src.parsers", "src.utils", "alembic", "alembic.versions"]
include-package-data = true

[tool.setuptools.package-data]
"alembic" = ["*.mako", "*.ini"]
"alembic.versions" = ["*.py"]

# ============================================================================
# Black - Code Formatter
# ============================================================================
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | node_modules
  | frontend
)/
'''

# ============================================================================
# MyPy - Static Type Checker
# ============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start permissive, tighten later
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
pretty = true

# Ignore missing imports for third-party libraries without stubs
[[tool.mypy.overrides]]
module = [
    "lxml.*",
    "uvicorn.*",
]
ignore_missing_imports = true

# ============================================================================
# Pytest - Testing Framework
# ============================================================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# ============================================================================
# Coverage - Code Coverage
# ============================================================================
[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "setup.py",
    "*/site-packages/*",
    "*/__pycache__/*",
    "frontend/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

# ============================================================================
# Ruff - Fast Python Linter (alternative/complement to pylint)
# ============================================================================
[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [
    "frontend",
    "node_modules",
    ".git",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ============================================================================
# Bandit - Security Linter
# ============================================================================
[tool.bandit]
exclude_dirs = ["tests", "frontend", "node_modules", ".venv", "venv"]
skips = ["B101", "B601"]  # B101: assert_used, B601: paramiko_calls
severity = "medium"
